// src/models/PolizaResponse.ts
export interface PolizaResponse {
    idPoliza: number;
    numeroPoliza: string;
    cedulaAsegurado: string;
    montoAsegurado: number;
    fechaVencimiento: string;
    fechaEmision: string;
    prima: number;
    periodo: string;
    fechaInclusion: string;
    aseguradora: string;
    idTipoPoliza: number;
    tipoPoliza: string;
    idEstadoPoliza: number;
    estadoPoliza: string;
    idCobertura: number;
    cobertura: string;
    nombreAsegurado: string;
    apellido1Asegurado: string;
    apellido2Asegurado: string;
  }
  